+++ src/App.css
@@ -39,3 +39,6 @@ footer{
   display: block;
   box-shadow: 0px -6px 9px 3px #cfcfcf;
 }
+.custom-modal {
+    width: 50%;
+}
