+++ src/App.js
@@ -3,8 +3,12 @@ import { Navbar, Nav, NavDropdown, NavItem, MenuItem } from 'react-bootstrap';
 import { bootstrapUtils } from 'react-bootstrap/lib/utils';
 import { LinkContainer } from 'react-router-bootstrap';
 import { observer } from 'mobx-react';
+import { FormGroup, FormControl, Modal, Button } from 'react-bootstrap';
+import ReactDOM from 'react-dom';
+import ajax from 'superagent';
 
 import LoginStatusUI from './components/LoginStatusUI';
+import baseUrl from './pages/config';
 import './App.css';
 
 bootstrapUtils.addStyle(Navbar, 'custom');
@@ -40,6 +44,52 @@ class Footer extends React.Component{
 
 @observer
 class App extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.state = {
+            showModal: false,
+            sugPlaceholder: "写下你的建议...",
+            sugValidationState: null
+        }
+    }
+
+    publish = () => {
+        const sugMessage = ReactDOM.findDOMNode(this.refs.sugValue).value.trim();
+
+        if(!sugMessage) {
+            this.errorReminder();
+            return;
+        }
+
+        const content = {
+            content: sugMessage
+        }
+        ajax.post(`${baseUrl}/suggestion/`)
+            .send(content)
+            .end((error,response) => {
+                if(error || response.status !== 201) {
+                    console.log('source push error!');
+                    alert("发布失败，请稍后再试");
+                    this.deleteInputValue();
+                } else {
+                    console.log('yay got ' + JSON.stringify(response.body));
+                    alert("发布成功");
+                    this.deleteInputValue();
+                }
+            })
+    }
+
+    errorReminder() {
+        if(ReactDOM.findDOMNode(this.refs.sugValue).value.trim() === "") {
+            this.setState({ sugPlaceholder : "提交内容不能为空..." });
+            this.setState({ sugValidationState : "error" });
+        }
+    }
+
+    deleteInputValue = () => {
+        ReactDOM.findDOMNode(this.refs.sugValue).value = "";
+    }
 
   getChildContext = () => {
     return {
@@ -48,42 +98,68 @@ class App extends React.Component {
   }
 
   render() {
+    let close = () => this.setState({ showModal: false });
+
     return (
-      <div className="header">
-        <Navbar bsStyle="custom">
-          <Navbar.Header>
-            <Navbar.Brand>
-              <a href="#">ForShare</a>
-            </Navbar.Brand>
-          </Navbar.Header>
-          <Nav>
-            <NavDropdown title={"资源分享"} id="basic-nav-dropdown">
-              <LinkContainer to="/articlesourcelist">
-                <MenuItem eventKey={1}>原创文章</MenuItem>
+      <div>
+        <div className="header">
+          <Navbar bsStyle="custom">
+            <Navbar.Header>
+              <Navbar.Brand>
+                <a href="#">ForShare</a>
+              </Navbar.Brand>
+            </Navbar.Header>
+            <Nav>
+              <NavDropdown title={"资源分享"} id="basic-nav-dropdown">
+                <LinkContainer to="/articlesourcelist">
+                  <MenuItem eventKey={1}>原创文章</MenuItem>
+                </LinkContainer>
+                <LinkContainer to="/linksourcelist">
+                  <MenuItem eventKey={2}>链接分享</MenuItem>
+                </LinkContainer>
+              </NavDropdown>
+              <LinkContainer to="newlink" activeHref="active">
+                <NavItem>上传链接</NavItem>
               </LinkContainer>
-              <LinkContainer to="/linksourcelist">
-                <MenuItem eventKey={2}>链接分享</MenuItem>
+              <LinkContainer to="newarticle" activeHref="active">
+                <NavItem>写文章</NavItem>
               </LinkContainer>
-            </NavDropdown>
-            <LinkContainer to="newlink" activeHref="active">
-              <NavItem>上传链接</NavItem>
-            </LinkContainer>
-            <LinkContainer to="newarticle" activeHref="active">
-              <NavItem>写文章</NavItem>
-            </LinkContainer>
-            <LinkContainer to="suggestion" activeHref="active">
-              <NavItem>意见反馈</NavItem>
-            </LinkContainer>
-          </Nav>
-          <Nav pullRight>
-            <LoginStatusUI></LoginStatusUI>
-          </Nav>
-        </Navbar>
-        <div className="body">
-          { this.props.children }
+              <LinkContainer to="suggestion" activeHref="active" onClick={() => this.setState({ showModal: true})}>  
+                <NavItem>意见反馈</NavItem>
+              </LinkContainer>
+            </Nav>
+            <Nav pullRight>
+              <LoginStatusUI></LoginStatusUI>
+            </Nav>
+          </Navbar>
+          <div className="body">
+            { this.props.children }
+          </div>
+          <Footer />
+        </div>  
+        <div>
+          <Modal 
+            show={ this.state.showModal }
+            onHide={ close }
+            dialogClassName="custom-modal"
+          >
+                <Modal.Header closeButton>
+                    <Modal.Title>意见反馈</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>
+                    <form>
+                        <FormGroup bsSize="large" validationState={this.state.sugValidationState}>
+                            <FormControl componentClass="textarea" placeholder={this.state.sugPlaceholder} ref="sugValue"/>
+                        </FormGroup>
+                    </form>
+                </Modal.Body>    
+                <Modal.Footer>
+                    <Button bsStyle="danger" onClick={close}>取消</Button>
+                    <Button bsStyle="danger" onClick={this.publish}>确认</Button>
+                </Modal.Footer>    
+            </Modal> 
         </div>
-        <Footer />
-      </div>
+      </div>   
     );
   }
 }
