--- src/pages/Suggestion.css
@@ -1,3 +1,0 @@
-.custom-modal {
-    width: 50%;
-}
